requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2024.10
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.1.5
django-appconf==1.0.6
django-crispy-forms==2.3
django-daisy==1.0.13
django-herald==0.3.0
django-imagekit==5.0.0
django-jazzmin==3.0.1
django-multiselectfield==0.1.13
django-taggit==6.1.0
gunicorn==23.0.0
idna==3.10
jsonpickle==4.0.2
jwt==1.3.1
kombu==5.4.2
packaging==24.2
pilkit==3.0
pillow==11.1.0
prompt_toolkit==3.0.49
psycopg2==2.9.10
pycparser==2.22
python-dateutil==2.9.0.post0
python-decouple==3.8
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
